fixed #1164: 1. Add the render mode which uses VBO without VAO in CCParticleSystemQuad. 2. delete some unused codes for win32 port. 3. Add CC_REBIND_INDICES_BUFFER macro.

+++ cocos2dx/tileMap_parallax_nodes/CCTMXLayer.cpp
@@ -233,7 +233,7 @@ void CCTMXLayer::parseInternalProperties()
 void CCTMXLayer::setupTileSprite(CCSprite* sprite, CCPoint pos, unsigned int gid)
 {
 	sprite->setPosition(positionAt(pos));
-	sprite->setVertexZ(vertexZForPos(pos));
+	sprite->setVertexZ((float)vertexZForPos(pos));
 	sprite->setAnchorPoint(CCPointZero);
 	sprite->setOpacity(m_cOpacity);
